backend/user/requirements.txt
astroid==2.4.2
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.22
protobuf==3.14.0
pycodestyle==2.6.0
PyJWT==2.0.1
pylint==2.6.0
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
urllib3==1.26.3
uuid==1.30
Werkzeug==1.0.1
wrapt==1.12.1
Flask-Cors==3.0.10